PR4

+++ intbuffet/src/main/java/com/js/intbuffetproject/dao/OrderDAO.java
@@ -17,5 +17,5 @@ public interface OrderDAO {
 	
 	Order getOrderById(Long id);
 
-	void removeOrder(Integer id);
+	void removeOrder(Long id);
 }
